Move user groups to separate files
