+++ src/web/server/app/routes/api.ts
@@ -90,70 +90,4 @@ app.get(
   }
 );
 
-/**
- * Add a user to group, admin locked
- */
-app.post(
-  '/addUserGroup',
-  [isLoggedIn, isAdmin],
-  (req: express.Request, res: express.Response) => {
-    apiController.addUserGroup(req, res);
-  }
-);
-
-/**
- * Sends a list of groupIds the logged in user is in
- */
-app.get(
-  '/getUserGroups',
-  isLoggedIn,
-  async (req: express.Request, res: express.Response) => {
-    apiController.retrieveUserGroups(req, res);
-  }
-);
-
-/**
- * Sends a list of userIds in a group
- */
-app.get(
-  '/getGroupUsers/:groupId',
-  isLoggedIn,
-  async (req: express.Request, res: express.Response) => {
-    apiController.retrieveGroupUsers(req, res);
-  }
-);
-
-/**
- * Sends a list of cameraIds in a group, admin locked
- */
-app.get(
-  '/getGroupCameras/:groupId',
-  [isLoggedIn, isAdmin],
-  async (req: express.Request, res: express.Response) => {
-    apiController.retrieveGroupCameras(req, res);
-  }
-);
-
-/**
- * Sends a list of cameraIds in groups the user is in
- */
-app.get(
-  '/getUserGroupCameras',
-  isLoggedIn,
-  async (req: express.Request, res: express.Response) => {
-    apiController.retrieveUserGroupCameras(req, res);
-  }
-);
-
-/**
- * Sends a dictionary of cameraId:streamId for all cameras in groups the user is in
- */
-app.get(
-  '/getLiveGroupCameraStreamIds',
-  isLoggedIn,
-  async (req: express.Request, res: express.Response) => {
-    apiController.retrieveLiveGroupCameraStreamIds(req, res);
-  }
-);
-
 module.exports = app;
