+++ src/web/server/app/routes/index.ts
@@ -3,7 +3,7 @@ import express from 'express';
 const app = express();
 
 app.use('/auth', require('./auth'));
-app.use('/api', require('./api'));
+app.use('/api', require('./api'), require('./apiGroups'));
 app.use('/notif', require('./notif'));
 
 export const routes = app;
